Split index page so we have one file/component representing the game that is entirely free of server-side rendering.
